Added note a/b posting on Meetup mailing lists
Added note about the need to check with Meetup admins prior to posting on their mailing lists.  I've been trying to delete the extra line spacing before and after the "Go to relevant meetups" bullet, but it won't go away.  My GitHub Markdown language appears to be properly written and doesn't contain any extra spaces.  If you're comfortable with it, I think it's ok for now, as it will not cause any confusion among readers.  If not, we can investigate further.